Move playbook files and tests to their own package
