+++ main_test.go
@@ -1,256 +1,1 @@
 package main
-
-import (
-	"bytes"
-	"errors"
-	"log"
-	"os"
-	"testing"
-)
-
-const PlaybookFilename = "test-playbook.yml"
-const ManifestFilename = "test-manifest.yml"
-const PlaybookContents string = `---
-name: The Project 
-meta:
-  team: Project Devs
-  email: devs@project.com
-  slack: devs
-vars:
-  - version
-  - assets_version
-  - owner
-tasks:
-  - name: Deploy Postgres
-    manifests:
-      - test-manifest.yml
-      - test-manifest.yml
-  - name: Deploy Redis
-    manifests:
-      - test-manifest.yml
-      - test-manifest.yml
-  - name: Database Setup
-    pod_manifest: test-manifest.yml
-    wait_for:
-      - success
-    when: new_deployment
-  - name: Database Migration
-    pod_manifest: test-manifest.yml
-    wait_for:
-      - success
-  - name: Deploy Project
-    manifests:
-      - test-manifest.yml
-      - test-manifest.yml
-      - test-manifest.yml
-`
-
-var PlaybookBytes = []byte(PlaybookContents)
-
-const PlaybookContentsIncomplete = `---
-name: The Project 
-`
-
-var IncompletePlaybookBytes = []byte(PlaybookContentsIncomplete)
-
-func TestMain(m *testing.M) {
-	f, err := os.Create(PlaybookFilename)
-	if err != nil {
-		log.Fatalf("Failed to write mock test playbook: %s", err)
-	}
-	f.Write(PlaybookBytes)
-	f.Close()
-
-	f, err = os.Create(ManifestFilename)
-	if err != nil {
-		log.Fatalf("Failed to write mock test manifest: %s", err)
-	}
-	f.Close()
-
-	testresult := m.Run()
-	teardown()
-	os.Exit(testresult)
-}
-func teardown() {
-	os.Remove(PlaybookFilename)
-	os.Remove(ManifestFilename)
-}
-
-func TestReadPlaybookFromDisk(t *testing.T) {
-	playbook, err := ReadPlaybookFromDisk(PlaybookFilename)
-	if err != nil {
-		t.Error(err)
-	}
-	if !bytes.Equal(playbook, PlaybookBytes) {
-		t.Error(errors.New("Playbook read from disk differs from Playbook written to disk"))
-	}
-}
-
-func TestParsePlaybook(t *testing.T) {
-	var err error
-	ParsedPlaybook, err := ParsePlaybook(PlaybookBytes)
-	if err != nil {
-		t.Error(err)
-	}
-	if ParsedPlaybook.Name != "The Project" {
-		t.Error(errors.New("Parsed Playbook has incorrect Name field"))
-	}
-}
-
-func TestParsePlaybookMalformed(t *testing.T) {
-	_, err := ParsePlaybook([]byte("asdf"))
-	if err == nil {
-		t.Error(errors.New("Parsing asdf succeeded, expected failure"))
-	}
-}
-
-func TestParsePlaybookIncomplete(t *testing.T) {
-	_, err := ParsePlaybook(IncompletePlaybookBytes)
-	if err != nil {
-		t.Error(errors.New("Parsing well-formed, incomplete playbook failed, expected success"))
-	}
-}
-
-func TestValidatePlaybookPasses(t *testing.T) {
-	ValidTask1 := Task{
-		Name:      "task",
-		Manifests: []string{ManifestFilename},
-	}
-	ValidTask2 := Task{
-		Name:        "task",
-		PodManifest: ManifestFilename,
-	}
-	testcases := []struct {
-		scenario string
-		playbook Playbook
-	}{
-		{
-			"Validate Valid Playbook",
-			Playbook{
-				Name:  "playbook 1",
-				Tasks: []Task{ValidTask1, ValidTask2},
-			},
-		},
-	}
-	for _, testcase := range testcases {
-		playbook := testcase.playbook
-		err := playbook.Validate()
-		if err != nil {
-			t.Errorf("Scenario %s\nExpected: No error\nActual:\n%s", testcase.scenario, err.Error())
-		}
-	}
-}
-
-func TestValidatePlaybookFailures(t *testing.T) {
-	InvalidTask1 := Task{
-		Manifests: []string{ManifestFilename},
-	}
-	InvalidTask2 := Task{
-		Name: "task",
-	}
-
-	testcases := []struct {
-		scenario    string
-		playbook    Playbook
-		expectedErr string
-	}{
-		{
-			"Validate Empty Playbook",
-			Playbook{},
-			"Playbook missing required Name",
-		},
-		{
-			"Validate Playbook With Zero Tasks",
-			Playbook{
-				Name:  "playbook 1",
-				Tasks: []Task{},
-			},
-			"Playbook requires at least 1 task",
-		},
-		{
-			"Validate Playbook With Tasks Missing Names",
-			Playbook{
-				Name:  "playbook 1",
-				Tasks: []Task{InvalidTask1},
-			},
-			"Task missing required Name",
-		},
-		{
-			"Validate Playbook With Tasks Missing Manifests",
-			Playbook{
-				Name:  "playbook 1",
-				Tasks: []Task{InvalidTask2},
-			},
-			"Task requires at least one manifest or a pod manifest",
-		},
-	}
-
-	for _, testcase := range testcases {
-		playbook := testcase.playbook
-		err := playbook.Validate()
-		if testcase.expectedErr != err.Error() {
-			t.Errorf("Scenario %s\nExpected:\n%s\nActual:\n%s", testcase.scenario, testcase.expectedErr, err.Error())
-		}
-	}
-}
-
-func TestTaskManifestsPresentPasses(t *testing.T) {
-	testcases := []struct {
-		scenario string
-		task     Task
-	}{
-		{
-			"Task With Existing Manifests",
-			Task{
-				Name:      "task 2",
-				Manifests: []string{ManifestFilename},
-			},
-		},
-		{
-			"Task With Only Pod Manifest",
-			Task{
-				Name:        "task 2",
-				PodManifest: ManifestFilename,
-			},
-		},
-		{
-			"Task With Both Manifests And Pod Manifest",
-			Task{
-				Name:        "task 2",
-				PodManifest: ManifestFilename,
-				Manifests:   []string{ManifestFilename},
-			},
-		},
-	}
-	for _, testcase := range testcases {
-		task := testcase.task
-		err := task.ManifestsPresent()
-		if err != nil {
-			t.Errorf("Scenario %s\nExpected: Success\nActual:\n%s", testcase.scenario, err.Error())
-		}
-	}
-}
-func TestTaskManifestsPresentFailures(t *testing.T) {
-	testcases := []struct {
-		scenario string
-		task     Task
-	}{
-		{
-			"Task With Missing Manifests",
-			Task{
-				Name:      "task 1",
-				Manifests: []string{"pod0"},
-			},
-		},
-	}
-
-	for _, testcase := range testcases {
-		task := testcase.task
-		err := task.ManifestsPresent()
-		if err == nil {
-			t.Errorf("Scenario %s\nExpected: File does not exist\nActual: Success%s", testcase.scenario)
-		} else if !os.IsNotExist(err) { // it was the wrong error!
-			t.Errorf("Scenario %s\nExpected: File does not exist\nActual:\n%s", testcase.scenario, err.Error())
-		}
-	}
-}
